+++ packages/amplify-category-api/src/graphql-transformer/transformer-options-v2.ts
@@ -291,6 +291,7 @@ const generateTransformParameters = (
       pathManager.getCurrentCloudBackendDirPath(),
     ),
     sandboxModeEnabled,
+    enableTransformerCfnOutputs: true,
   };
 };