+++ packages/amplify-graphql-api-construct/src/__tests__/__functional__/relationships.test.ts
@@ -1,7 +1,7 @@
 import * as cdk from 'aws-cdk-lib';
 import { Template } from 'aws-cdk-lib/assertions';
 import { AmplifyGraphqlApi } from '../../amplify-graphql-api';
-import { AmplifyGraphqlSchema } from '../../amplify-graphql-schema';
+import { AmplifyGraphqlDefinition } from '../../amplify-graphql-definition';
 
 /**
  * Utility to wrap construct creation a basic synth step to smoke test
@@ -10,8 +10,8 @@ import { AmplifyGraphqlSchema } from '../../amplify-graphql-schema';
 const verifySchema = (schema: string): void => {
   const stack = new cdk.Stack();
   new AmplifyGraphqlApi(stack, 'TestApi', {
-    schema: AmplifyGraphqlSchema.fromString(schema),
-    authorizationConfig: {
+    definition: AmplifyGraphqlDefinition.fromString(schema),
+    authorizationModes: {
       apiKeyConfig: { expires: cdk.Duration.days(7) },
     },
   });