+++ packages/amplify-graphql-api-construct/src/__tests__/__functional__/resources.test.ts
@@ -1,7 +1,7 @@
 import * as cdk from 'aws-cdk-lib';
 import * as cognito from 'aws-cdk-lib/aws-cognito';
 import { AmplifyGraphqlApi } from '../../amplify-graphql-api';
-import { AmplifyGraphqlSchema } from '../../amplify-graphql-schema';
+import { AmplifyGraphqlDefinition } from '../../amplify-graphql-definition';
 
 describe('generated resource access', () => {
   describe('l1 resources', () => {
@@ -14,12 +14,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnGraphqlApi, cfnGraphqlSchema, cfnApiKey },
           },
         } = new AmplifyGraphqlApi(stack, 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: owner }]) {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             userPoolConfig: { userPool },
           },
         });
@@ -35,12 +35,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnApiKey },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -56,12 +56,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnResolvers },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -83,12 +83,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnResolvers },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) @searchable {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -103,12 +103,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnResolvers },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String! @index
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -123,7 +123,7 @@ describe('generated resource access', () => {
             cfnResources: { cfnResolvers },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String!
               authors: [Author] @hasMany
@@ -145,7 +145,7 @@ describe('generated resource access', () => {
               todo: Todo @belongsTo
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -187,12 +187,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnDataSources },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -208,12 +208,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnDataSources },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) @searchable {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -228,12 +228,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnDataSources },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Query {
               echo(message: String!): String! @function(name: "echo")
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -250,13 +250,16 @@ describe('generated resource access', () => {
             cfnResources: { cfnFunctionConfigurations },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
+          },
+          translationBehavior: {
+            disableResolverDeduping: false,
           },
         });
 
@@ -287,7 +290,7 @@ describe('generated resource access', () => {
             cfnResources: { cfnTables },
           },
         } = new AmplifyGraphqlApi(stack, 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: owner }]) {
               description: String!
             }
@@ -302,7 +305,7 @@ describe('generated resource access', () => {
               posts: [Post] @manyToMany(relationName: "PostAuthors")
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             userPoolConfig: { userPool },
           },
         });
@@ -322,7 +325,7 @@ describe('generated resource access', () => {
             cfnResources: { cfnTables },
           },
         } = new AmplifyGraphqlApi(stack, 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: owner }]) {
               description: String!
             }
@@ -333,7 +336,7 @@ describe('generated resource access', () => {
               detectionType: 'VERSION',
             },
           },
-          authorizationConfig: {
+          authorizationModes: {
             userPoolConfig: { userPool },
           },
         });
@@ -351,12 +354,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnFunctions },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) @searchable {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -373,12 +376,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnRoles },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -393,12 +396,12 @@ describe('generated resource access', () => {
             cfnResources: { cfnRoles },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) @searchable {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -417,12 +420,12 @@ describe('generated resource access', () => {
             cfnResources: { additionalCfnResources },
           },
         } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-          schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Todo @model @auth(rules: [{ allow: public }]) @searchable {
               description: String!
             }
           `),
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         });
@@ -439,12 +442,12 @@ describe('generated resource access', () => {
       const {
         resources: { graphqlApi },
       } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-        schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+        definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
           type Todo @model @auth(rules: [{ allow: public }]) {
             description: String!
           }
         `),
-        authorizationConfig: {
+        authorizationModes: {
           apiKeyConfig: { expires: cdk.Duration.days(7) },
         },
       });
@@ -456,12 +459,12 @@ describe('generated resource access', () => {
       const {
         resources: { tables },
       } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-        schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+        definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
           type Todo @model @auth(rules: [{ allow: public }]) {
             description: String!
           }
         `),
-        authorizationConfig: {
+        authorizationModes: {
           apiKeyConfig: { expires: cdk.Duration.days(7) },
         },
       });
@@ -474,12 +477,12 @@ describe('generated resource access', () => {
       const {
         resources: { roles },
       } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-        schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+        definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
           type Todo @model @auth(rules: [{ allow: public }]) {
             description: String!
           }
         `),
-        authorizationConfig: {
+        authorizationModes: {
           apiKeyConfig: { expires: cdk.Duration.days(7) },
         },
       });
@@ -492,12 +495,12 @@ describe('generated resource access', () => {
       const {
         resources: { functions },
       } = new AmplifyGraphqlApi(new cdk.Stack(), 'TestApi', {
-        schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+        definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
           type Todo @model @auth(rules: [{ allow: public }]) @searchable {
             description: String!
           }
         `),
-        authorizationConfig: {
+        authorizationModes: {
           apiKeyConfig: { expires: cdk.Duration.days(7) },
         },
       });